requirements.txt
SQLAlchemy
python-telegram-bot